requirements.txt
aiohttp
numpy
pandas
python-dotenv
pytest
pytest-asyncio
black
flake8
mypy
